Reorganization update.
API is now its own process. Updates via nodemon just like the web
server. Various project components have been separated and reorganized
to hopefully make it more clear what files go with which part of the
application.

For instance, the new /web/ folder has two sub-directories for /client/
and /server/ rendered web files./components.

Servers are now in their own directory. The web and api HAPI servers
should be easy to identify.

The /api/ directory has all the api handlers and routes in it.

For now this is a better setup for me, but I may make further
refinements. Moving the API and WEB servers into their own packages
might be worthwhile.

In the meantime I'm going to expand the API and tests to include more
algorithms.