Re-arranged project. Again
Now API and WEB have strong separation. In fact, each is now a separate
package and can be run/moved/whatever entirely independently. This'll
set up a neat problem for later when I split the APIs and wire-up the
web project to pull data from many different ports/processes/etc.

It'll start to make more sense once I get a database hooked up, but now
that API (at least) is on it's own I can use it as a template for
spinning up new microservices quickly.