Re-organization and clean up.
Ok, still not sure if I'm satisfied.

For one, it would be nice to have the paths be
/fibonacci/components and /fibonacci/combiners which doesn't affect
depth but groups things a bit better, but that leaves a need for a
/components (and maaaaaybe a /combiners.... So I'll leave it. For now.

I'd also like to find a way for components to name their own store
objects... I know it's not really "the redux way" but being able to use
a component (from a page level) and not have to go re-wire the store
WOULD be kinda nice...

If I can get the Nth and UpToNth components/combiners running on the
same page and operating independently with arbitrary named instances in
the store.fibonacci object I'll call it a success.